Implement premium waiver. In the contract class, when the premiums are waived, all future cash flows and derived preent values from that time on are recalculated. The premium-free sum insured is already known, as that is one column from the original reserves table.
Added flags to indicate that no further surrender penalty should be applied. Flag for alpha-cost refund has been added, but is no used yet.

In the premium waiver case, the premium cashflows from t onwards are set to 0, the gamma costs are set to gamma_nopremiums and no surrender penalty is applied any more. Apart from that, the present value and reserve calculation is exactly the same as in the original contract.