Implement insurances on multiple lives that have one joint mortality table
The only extension for joint-lives with one joint life table is to allow the age to be a vector and store the ageDifferences in the ContractData. The mortalityTable then needs to be a life table that picks up the age differences and returns a proper joint life table.

By default, if age is a vector, the technical age of the contract will be the first vector, so in many cases one does not even need to override the technicalAge parameter.