Fix dynamics that start at t=0
Problem was that padding convert the date to a numeric, so summing the date from the original main contract part with any other dynamic block worked. But when the dynamic block was not padded, because it started at t=0, the date was not (wrongly) converted to numeric, so R failed adding two dates...

For the balance sheet reserve, I now simply match corresponding rows by date (available as a column in the data.frame) rather than relying on information how many rows to pad.

closes #65