Add unitcostsInGross Feature flag
By default, the gross premium does not include unit costs, which are added after gross premium (but before taxes). Some companies include the unitcosts in the gross premium (which is refunded upon death -> correct calculation is essentaion). This flag allows a tarif to include the unit cost in the gross premium reather than adding later.

Fixes Issue #73 (except for the documentation, which is covered by a different issue anyway)