Add profitAttributionAtExpiration flag
If this is set to TRUE, a profit will be calculated at the expiration date (if it is an anniversary). When the profit cycle does not use immediate profit attribution, no profit shall be declared at expiration, since there won't be a date to attribute it to the contract.

Also, don't pass the waiting vector to the terminal bonus calculation. The terminal bonus is typically not affected by waiting periods for profit participation (and the profitAttributionAtExpiration  flag mis-uses the waiting vector to prevent profit calculation at the expiration date).